requirements.txt
# Instagram Warm-Up Automation
# Python 3.9+

# Appium client for Android automation
Appium-Python-Client>=3.0.0

# Optional: UIAutomator2 backend (often used with Appium for Android)
# uiautomator2>=2.16.0

# Config (YAML)
PyYAML>=6.0

# State and config
# (sqlite3 is stdlib; no extra deps for JSON)
